+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # monopro-ai-cli
 
+## 0.4.0
+
+### Minor Changes
+
+- Beta use of monopro services and CLI
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - monopro-ai@0.4.0
+
+## 0.3.0
+
+### Minor Changes
+
+- Beta use of monopro services and CLI
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - monopro-ai@0.3.0
+
 ## 0.2.1
 
 ### Patch Changes